requirements.txt
streamlit
openai
PyGithub
faiss-cpu
sentence-transformers
tree-sitter-python
